requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
tinydb==4.8.2
Werkzeug==3.1.3
python-dotenv
azure-cosmos
azure-monitor-opentelemetry
annotated-types==0.7.0
anyio==4.13.0
asarPy==1.0.1
certifi==2026.2.25
distro==1.9.0
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
